Clean the array in minimum number of operations.
Geek has been given an array arr[] of size n by his teacher. His task is to make the array empty using a specific operation. In each operation, Geek can select a subsequence from the array such that all elements in that subsequence are distinct. The goal is to empty the array using the minimum number of such operations.