To check if List of Strings are Pangram or not.
A string is a pangram if it contains all letters of the English alphabet, ascii['a'-'z']. Given a list of strings, determine if each one is a pangram or not. Return "1" if true and "0" if false.
Example:
pangram = ['pack my box with five dozen liquor jugs', 'this is not a pangram']
the string pack my box with five dozen liquor jugs is a pangram, because it contains all the letters 'a' through 'z'
the string 'this is not a pangram' is not a pangram
Assemble a string of the two results, in order. The result is '10'.