Find the minimum possible length substring.
Given a string, str, that consists of the characters 'A' and 'B' only, in one move, delete either an "AB" or a "BB" substring and concatenate the remaining substrings. Find the minimum possible length of the remaining string after performing any number of moves. Note: A substring is a contiguous subsequence of a string. Example seq = "BABBA" Using 0-based indexing, the following moves are optimal. * Delete the substring "AB" starting at index 1. "BABBA" "BBA" Delete the substring "BB" starting at index 0. "BBA" to "A".